+++ src/options.js
@@ -3,9 +3,11 @@
  *
  * @type {Object}
  * @property {boolean} excludeSymlinks Whether to exclude symbolic links or not.
+ * @property {boolean} directoriesFirst Whether to list directories before files.
  */
 const defaults = {
   excludeSymlinks: false,
+  directoriesFirst: false,
 };
 
 export default defaults;