+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file. The format 
 
 ## [Unreleased]
 
-(nothing)
+- `directoriesFirst` option - [#8]
 
 ## [v0.2.0] - 2018-11-23
 
@@ -47,6 +47,7 @@ All notable changes to this project will be documented in this file. The format 
 [#4]: https://github.com/amercier/files-by-directory/pull/4
 [#5]: https://github.com/amercier/files-by-directory/pull/5
 [#7]: https://github.com/amercier/files-by-directory/pull/7
+[#8]: https://github.com/amercier/files-by-directory/pull/8
 [v0.1.0]: https://github.com/amercier/files-by-directory/compare/v0.0.0...v0.1.0
 [v0.1.1]: https://github.com/amercier/files-by-directory/compare/v0.1.0...v0.1.1
 [v0.1.2]: https://github.com/amercier/files-by-directory/compare/v0.1.1...v0.1.2